trocado valor de long para double

+++ src/main/java/entidades/ProdutoDigital.java
@@ -3,7 +3,7 @@ package entidades;
 public class ProdutoDigital extends Produto{
     private Long tamanhoProgramaBytes;
 
-    public ProdutoDigital(Float preco, String nome, Integer codigoProduto, Long tamanhoProgramaBytes) {
+    public ProdutoDigital(Double preco, String nome, Integer codigoProduto, Long tamanhoProgramaBytes) {
         super(preco, nome, codigoProduto);
         this.tamanhoProgramaBytes = tamanhoProgramaBytes;
     }
